Moved the patched Birch class from exterals to fd.cluster
