+++ freediscovery/cluster/__init__.py
@@ -2,6 +2,7 @@
 
 from .base import (select_top_words, ClusterLabels)
 from .utils import centroid_similarity
+from .birch import Birch
 from .optimal_sampling import compute_optimal_sampling
 
 from .dendrogram import _DendrogramChildren
